Stop OSSAR scanning
The OSSAR code scanning only works with public repositories. As the repository has been changed to private, all the statements has been commented to prevent scanning attempts. The file has not been deleted just as is not sure if the status is going to change to public in the future.